Split PSX RIC
