+++ config/splat.hd.ric.yaml
@@ -50,8 +50,8 @@ segments:
       - [0x18F7C, .data, pl_blueprints]
       - [0x1938C, .data, pl_anims]
       - [0x199AC, .data, pl_whip]
-      - [0x19D64, .data, 2BA68]
-      - [0x19D94, .data, 2C4C4]
+      - [0x19D64, .data, pl_subweapon_holywater]
+      - [0x19DDC, .data, pl_subweapon_cross]
       - [0x19E70, .data, pl_subweapons]
       - [0x19E98, .data, 2F8E8]
       - [0x19EE0, .data, 319C4]
@@ -64,7 +64,8 @@ segments:
       - [0x1A8D0, .rodata, pl_collision]
       - [0x1A8D4, .rodata, pl_blueprints]
       - [0x1AA6C, .rodata, pl_whip]
-      - [0x1AA88, .rodata, 2C4C4]
+      - [0x1AA88, .rodata, pl_subweapon_holywater]
+      - [0x1AAA0, .rodata, pl_subweapon_cross]
       - [0x1AAF0, .rodata, 2F8E8]
       - [0x1AB04, .rodata, 319C4]
       - [0x1AC1C, .rodata, pl_crash_cross]
@@ -77,8 +78,8 @@ segments:
       - [0x23834, c, pl_blueprints]
       - [0x2A6C0, c, pl_whip]
       - [0x2B994, c, pl_subweapon_dummy]
-      - [0x2B9A4, c, 2BA68]
-      - [0x2C42C, c, 2C4C4]
+      - [0x2B9A4, c, pl_subweapon_holywater]
+      - [0x2CF80, c, pl_subweapon_cross]
       - [0x2E1C0, c, pl_subweapons]
       - [0x2F83C, c, 2F8E8]
       - [0x31918, c, 319C4]
@@ -87,7 +88,7 @@ segments:
       - [0x38ED0, .bss, pl_collision]
       - [0x38ED4, .bss, pl_blueprints]
       - [0x38FD4, .bss, pl_whip]
-      - [0x38FDC, .bss, 2C4C4]
+      - [0x38FDC, .bss, pl_subweapon_cross]
       - [0x397E0, .bss, pl_subweapons]
       - [0x397E4, .bss, 2F8E8]
       - [0x397E8, .bss, 319C4]
