+++ src/ric/pl_subweapon_cross.c
@@ -2,316 +2,6 @@
 #include "ric.h"
 #include "sfx.h"
 
-// Unused?
-static s16 D_80155D94[] = {16, 0, -1, 0};
-
-// Entity ID #7. Made by blueprint 6. Comes from subweapon 3. Holy water!
-// Not at all the same as DRA's.
-void RicEntitySubwpnHolyWater(Entity* self) {
-    s16 argY;
-    s32 collision_result;
-    s32 tempXVel;
-    s32 trigresult;
-    s32 trigtemp;
-
-    if (self->step >= 3) {
-        self->posY.i.hi += 5;
-    }
-    switch (self->step) {
-    case 0:
-        self->flags = FLAG_POS_CAMERA_LOCKED;
-        self->animSet = ANIMSET_OVL(0x12);
-        self->animCurFrame = 0x23;
-        self->zPriority = PLAYER.zPriority + 2;
-        self->unk5A = 0x46;
-        self->palette = PAL_OVL(0x12F);
-        self->posX.i.hi = self->posX.i.hi;
-        self->posY.i.hi = self->posY.i.hi - 0x10;
-        self->ext.holywater.angle = (rand() & 0x7F) + 0xDC0;
-        if (PLAYER.facingLeft == 1) {
-            self->ext.holywater.angle = (rand() & 0x7F) + 0x9C0;
-        }
-
-        trigresult = rcos(self->ext.holywater.angle);
-        trigtemp = trigresult * 16;
-        self->velocityX = (trigresult * 32 + trigtemp) << 9 >> 8;
-
-        trigresult = rsin(self->ext.holywater.angle);
-        trigtemp = trigresult * 16;
-        self->velocityY = -((trigresult * 32 + trigtemp) << 9) >> 8;
-
-        self->ext.holywater.subweaponId = PL_W_HOLYWATER;
-        RicSetSubweaponParams(self);
-        self->hitboxWidth = 4;
-        self->hitboxHeight = 4;
-        self->ext.holywater.unk80 = 0x200;
-        self->step += 1;
-        break;
-    case 1:
-        self->posY.val += self->velocityY;
-        collision_result = RicCheckHolyWaterCollision(0, 0);
-        self->posX.val += self->velocityX;
-
-        if (self->velocityX < 0) {
-            argY = -4;
-        } else {
-            argY = 4;
-        }
-        collision_result |= func_8016840C(-7, argY);
-        if (collision_result & 2) {
-            self->velocityX = -self->velocityX;
-            collision_result = 1;
-        }
-        if ((collision_result & 1) || (self->hitFlags != 0)) {
-            g_api.CreateEntFactoryFromEntity(self, 0x28U, 0);
-            g_api.PlaySfx(SFX_GLASS_BREAK_WHOOSH);
-            self->ext.holywater.timer = 0x50;
-            self->animSet = 0;
-            self->hitboxState = 0;
-            self->step = 3;
-            self->velocityX >>= 2;
-        }
-        break;
-    case 2:
-        if (--self->ext.holywater.timer == 0) {
-            self->ext.holywater.timer = 0x50;
-            self->velocityX >>= 2;
-            self->step++;
-        }
-        break;
-    case 3:
-        if (!(self->ext.holywater.timer & 3)) {
-            RicCreateEntFactoryFromEntity(
-                self, FACTORY(BP_HOLYWATER_FIRE, self->ext.holywater.unk82), 0);
-            // weirdly nonmatching with ++;
-            self->ext.holywater.unk82 = self->ext.holywater.unk82 + 1;
-            self->velocityX -= (self->velocityX / 32);
-        }
-        collision_result = RicCheckHolyWaterCollision(6, 0);
-        tempXVel = self->velocityX;
-        if (self->velocityX < 0) {
-            if ((collision_result & 0xF000) == 0xC000) {
-                tempXVel = tempXVel * 10 / 16;
-            }
-            if ((collision_result & 0xF000) == 0xD000) {
-                tempXVel = tempXVel * 13 / 16;
-            }
-        }
-        if (self->velocityX > 0) {
-            if ((collision_result & 0xF000) == 0x8000) {
-                tempXVel = tempXVel * 10 / 16;
-            }
-            if ((collision_result & 0xF000) == 0x9000) {
-                tempXVel = tempXVel * 13 / 16;
-            }
-        }
-        self->posX.val += tempXVel;
-        if (self->velocityX < 0) {
-            argY = -4;
-        } else {
-            argY = 4;
-        }
-        collision_result |= func_8016840C(-7, argY);
-        if (!(collision_result & 1)) {
-            self->velocityX >>= 1;
-            self->step++;
-        }
-        break;
-    case 4:
-        if (!(self->ext.holywater.timer & 3)) {
-            RicCreateEntFactoryFromEntity(
-                self, FACTORY(BP_HOLYWATER_FIRE, self->ext.holywater.unk82), 0);
-            self->ext.holywater.unk82 = self->ext.holywater.unk82 + 1;
-        }
-        self->velocityY += FIX(12.0 / 128);
-        if (self->velocityY > FIX(4)) {
-            self->velocityY = FIX(4);
-        }
-        self->posY.val += self->velocityY;
-        collision_result = RicCheckHolyWaterCollision(0, 0);
-        self->posX.val += self->velocityX;
-        if (self->velocityX < 0) {
-            argY = -4;
-        } else {
-            argY = 4;
-        }
-        collision_result |= func_8016840C(-7, argY);
-        if (collision_result & 1) {
-            self->velocityX *= 2;
-            self->step--;
-        }
-        break;
-    case 5:
-        break;
-    }
-    if (self->step >= 3) {
-        if (--self->ext.holywater.timer < 0) {
-            DestroyEntity(self);
-            return;
-        }
-        if (self->ext.holywater.timer == 2) {
-            self->step = 5;
-        }
-        self->animCurFrame = 0;
-        self->hitboxState = 0;
-        self->posY.i.hi -= 5;
-    }
-    g_Player.timers[PL_T_3] = 2;
-}
-
-// Entity ID #8. Blueprint 7.
-static Point32 D_80155D9C[] = {{28, 0}, {28, 16}, {28, 32}, {28, 48},
-                               {60, 0}, {60, 16}, {60, 32}, {60, 48}};
-void RicEntitySubwpnHolyWaterFlame(Entity* self) {
-    s16 sp10[5];
-    s16 sp20[5];
-    s16 pad[2];
-    Primitive* prim;
-    s16 angleTemp;
-    s32 angle;
-    s32 i;
-    s32 hex80;
-    u8 randR;
-    u8 randG;
-    u8 randB;
-    u8 primUBase;
-    u8 primVBase;
-    s16* primYPtr;
-    s32 var_s4;
-    s16 temp_v0_2;
-    s32 doubleparams;
-
-    s16 upperParams = self->params >> 8;
-
-    primUBase = D_80155D9C[(g_GameTimer & 7)].x;
-    primVBase = D_80155D9C[(g_GameTimer & 7)].y;
-    switch (self->step) {
-    case 0:
-        randR = (rand() & 0x3F) + 0x2F;
-        randG = (rand() & 0x3F) + 0x6F;
-        randB = (rand() & 0x7F) + 0x6F;
-        self->primIndex = g_api.AllocPrimitives(PRIM_GT4, 4);
-        if (self->primIndex == -1) {
-            DestroyEntity(self);
-            return;
-        }
-        prim = &g_PrimBuf[self->primIndex];
-        i = 0;
-        while (prim != NULL) {
-            PRED(prim) = randR;
-            PGRN(prim) = randG;
-            PBLU(prim) = randB;
-            if (i == 0) {
-                prim->b0 = prim->b1 = prim->g0 = prim->g1 = prim->r0 =
-                    prim->r1 = 0;
-            }
-            prim->clut = 0x1B0;
-            prim->tpage = 0x1A;
-            prim->priority = PLAYER.zPriority + 2;
-            prim->drawMode = DRAW_UNK_40 | DRAW_TPAGE2 | DRAW_TPAGE |
-                             DRAW_HIDE | DRAW_COLORS | DRAW_UNK02 | DRAW_TRANSP;
-            prim = prim->next;
-            i++;
-        }
-        self->flags = FLAG_POS_CAMERA_LOCKED | FLAG_HAS_PRIMS;
-        self->ext.holywater.timer = 1;
-        self->step += 1;
-        break;
-    case 1:
-        if (--self->ext.holywater.timer == 0) {
-            self->ext.holywater.subweaponId = PL_W_HOLYWATER_FLAMES;
-            RicSetSubweaponParams(self);
-            self->hitboxWidth = 4;
-            self->ext.holywater.hitboxState = self->hitboxState;
-            self->posY.i.hi = self->posY.i.hi - 10;
-            RicCreateEntFactoryFromEntity(self, FACTORY(BP_EMBERS, 3), 0);
-            self->ext.holywater.timer = 0x50;
-            self->posY.i.hi = self->posY.i.hi + 10;
-            self->ext.holywater.unk80 = (rand() & 0xF) + 0x12;
-            self->ext.holywater.angle = rand() & 0xFFF;
-            self->step += 1;
-        }
-        break;
-    case 2:
-        if (self->facingLeft) {
-            var_s4 = 1;
-        } else {
-            var_s4 = -1;
-        }
-        angleTemp = self->ext.holywater.angle;
-        self->ext.holywater.angle += 0xC0;
-        angle = angleTemp;
-        for (i = 0; i < 4; i++) {
-            sp10[i] = self->posX.i.hi + (rsin(angle) >> 0xA);
-            angle += 0x400;
-        }
-        sp10[4] = self->posX.i.hi;
-        sp10[0] = var_s4 + self->posX.i.hi;
-        temp_v0_2 = rsin((s16)((self->ext.holywater.timer * 64) + 0x800)) >> 8;
-        temp_v0_2 += self->ext.holywater.unk80;
-        temp_v0_2 = temp_v0_2 * 3 >> 1;
-        sp20[0] = self->posY.i.hi - temp_v0_2;
-        sp20[4] = self->posY.i.hi;
-        sp20[2] = (sp20[0] + sp20[4]) / 2;
-        sp20[1] = (sp20[0] + sp20[2]) / 2;
-        sp20[3] = (sp20[2] + sp20[4]) / 2;
-        prim = &g_PrimBuf[self->primIndex];
-        if (--self->ext.holywater.timer < 0) {
-            DestroyEntity(self);
-            return;
-        }
-        if (self->ext.holywater.timer & 3) {
-            self->hitboxState = 0;
-        } else {
-            self->hitboxState = self->ext.holywater.hitboxState;
-        }
-        if (self->ext.holywater.timer < 0x15) {
-            self->hitboxState = 0;
-        }
-        i = 0;
-        while (prim != NULL) {
-            if (upperParams * 2 + 0x18 >= self->ext.holywater.timer) {
-                if (prim->g0 >= 10) {
-                    prim->g0 -= 5;
-                }
-                if (prim->b0 >= 10) {
-                    prim->b0 -= 5;
-                }
-                if (prim->r0 >= 16) {
-                    prim->r0 -= 8;
-                }
-                prim->g1 = prim->g2 = prim->g3 = prim->g0;
-                prim->r1 = prim->r2 = prim->r3 = prim->r0;
-                prim->b1 = prim->b2 = prim->b3 = prim->b0;
-            }
-            prim->x0 = sp10[i] - 8;
-            prim->x1 = sp10[i] + 8;
-            prim->y0 = sp20[i];
-            prim->y1 = sp20[i];
-            prim->x2 = sp10[i + 1] - 8;
-            prim->x3 = sp10[i + 1] + 8;
-            prim->y2 = sp20[i + 1];
-            prim->y3 = sp20[i + 1];
-            prim->drawMode &= ~DRAW_HIDE;
-            hex80 = 0x80;
-            prim->u0 = prim->u1 = primUBase - ((i * 7) + hex80);
-            prim->u2 = prim->u3 = primUBase - (((i + 1) * 7) + hex80);
-            prim->v0 = prim->v2 = primVBase - hex80;
-            prim->v1 = prim->v3 = primVBase - 0x70;
-            if ((sp20[4] - sp20[0]) < 7) {
-                prim->drawMode |= DRAW_HIDE;
-                DestroyEntity(self);
-                return;
-            }
-            i++;
-            prim = prim->next;
-        }
-        self->hitboxHeight = temp_v0_2 >> 1;
-        self->hitboxOffY = (-temp_v0_2 >> 1);
-    }
-    g_Player.timers[PL_T_3] = 2;
-}
-
 // Entity 13. Made by blueprint 13. That's from subweapon 12.
 // That's the crash for subweapon 4. That's the cross.
 static u16 crash_cross_img_data[] = {
