+++ src/ric/pl_subweapon_holywater.c
@@ -1,0 +1,605 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "ric.h"
+#include "sfx.h"
+
+void RicEntityNotImplemented3(Entity* self) {}
+
+// Entity ID #35. Created by blueprint 40. No known FACTORY calls with
+// blueprint 40. Duplicate of DRA EntityHolyWaterBreakGlass.
+s16 D_80155D64[4][6] = {
+    {2, -2, 0, -4, 0, 0},
+    {-3, -3, -1, 1, 2, 0},
+    {-4, -3, 2, -2, -2, 2},
+    {-1, 0, 0, -4, 3, 3},
+};
+extern s16 D_80155D64[][6];
+void func_80167A70(Entity* self) {
+    Point16 sp10[8];
+    Primitive* prim;
+    FakePrim* fakeprim;
+    s32 velX;
+    s32 i;
+    u16 posY;
+    u16 posX;
+    u8 arrIndex;
+    u8 randbit;
+
+    switch (self->step) {
+    case 0:
+        self->primIndex = g_api.AllocPrimitives(PRIM_GT4, 16);
+        if (self->primIndex == -1) {
+            DestroyEntity(self);
+            return;
+        }
+
+        prim = &g_PrimBuf[self->primIndex];
+        posX = self->posX.i.hi;
+        posY = self->posY.i.hi;
+
+        for (i = 0; prim != NULL; i++, prim = prim->next) {
+            if (i < 8) {
+                fakeprim = (FakePrim*)prim;
+                fakeprim->x0 = posX;
+                fakeprim->posX.i.hi = posX;
+                sp10[i].x = posX;
+                fakeprim->y0 = posY;
+                fakeprim->posY.i.hi = posY;
+                sp10[i].y = posY;
+                // Random velocity from 0.25 to 0.5
+                velX = (rand() & 0x3FFF) + FIX(0.25);
+                fakeprim->velocityX.val = velX;
+                if (i & 1) {
+                    fakeprim->velocityX.val = -velX;
+                }
+                fakeprim->velocityY.val = -((rand() * 2) + FIX(2.5));
+                fakeprim->drawMode = DRAW_HIDE | DRAW_UNK02;
+                fakeprim->type = PRIM_TILE;
+            } else {
+                prim->r0 = prim->r1 = prim->r2 = prim->r3 =
+                    (rand() & 0xF) | 0x30;
+                prim->b0 = prim->b1 = prim->b2 = prim->b3 = rand() | 0x80;
+                prim->g0 = prim->g1 = prim->g2 = prim->g3 =
+                    (rand() & 0x1F) + 0x30;
+                randbit = rand() & 1;
+                prim->drawMode = !(randbit) ? 6 : 0x37;
+                posX = sp10[i - 8].x;
+                posY = sp10[i - 8].y;
+                arrIndex = i & 3;
+                prim->u0 = arrIndex;
+                prim->x0 = posX + D_80155D64[arrIndex][0];
+                prim->y0 = posY + D_80155D64[arrIndex][1];
+                prim->x1 = posX + D_80155D64[arrIndex][2];
+                prim->y1 = posY + D_80155D64[arrIndex][3];
+                prim->x3 = prim->x2 = posX + D_80155D64[arrIndex][4];
+                prim->y3 = prim->y2 = posY + D_80155D64[arrIndex][5];
+                prim->type = PRIM_G4;
+                prim->priority = PLAYER.zPriority + 2;
+            }
+        }
+        self->flags = FLAG_POS_CAMERA_LOCKED | FLAG_HAS_PRIMS;
+        self->ext.timer.t = 20;
+        self->step++;
+        break;
+
+    case 1:
+        if (--self->ext.timer.t == 0) {
+            DestroyEntity(self);
+            return;
+        }
+
+        prim = &g_PrimBuf[self->primIndex];
+        for (i = 0; prim != NULL; i++, prim = prim->next) {
+            if (i < 8) {
+                fakeprim = (FakePrim*)prim;
+                fakeprim->posX.i.hi = fakeprim->x0;
+                fakeprim->posY.i.hi = fakeprim->y0;
+                fakeprim->posX.val += fakeprim->velocityX.val;
+                fakeprim->posY.val += fakeprim->velocityY.val;
+                fakeprim->velocityY.val += FIX(36.0 / 128);
+                sp10[i].x = fakeprim->posX.i.hi;
+                sp10[i].y = fakeprim->posY.i.hi;
+                fakeprim->x0 = fakeprim->posX.i.hi;
+                fakeprim->y0 = fakeprim->posY.i.hi;
+            } else {
+                posX = sp10[i - 8].x;
+                posY = sp10[i - 8].y;
+                arrIndex = prim->u0;
+                prim->x0 = posX + D_80155D64[arrIndex][0];
+                prim->y0 = posY + D_80155D64[arrIndex][1];
+                prim->x1 = posX + D_80155D64[arrIndex][2];
+                prim->y1 = posY + D_80155D64[arrIndex][3];
+                prim->x3 = prim->x2 = posX + D_80155D64[arrIndex][4];
+                prim->y3 = prim->y2 = posY + D_80155D64[arrIndex][5];
+            }
+        }
+        break;
+    }
+}
+
+// Entity ID #11. Created by blueprint 12.
+// This is blueprintNum for subweapon ID 16.
+// That is the crash for subweapon 3. That's holy water!
+void RicEntityCrashHydroStorm(Entity* self) {
+    PrimLineG2* line;
+    s16 primcount;
+    s32 trigresult;
+    s32 trigtemp;
+
+    if (self->params < 40) {
+        primcount = 32;
+    } else {
+        primcount = 33 - ((self->params - 32) * 2);
+    }
+
+    switch (self->step) {
+    case 0:
+        self->primIndex = g_api.AllocPrimitives(PRIM_LINE_G2, primcount);
+        if (self->primIndex == -1) {
+            DestroyEntity(self);
+            return;
+        }
+        self->ext.subweapon.subweaponId = PL_W_HYDROSTORM;
+        RicSetSubweaponParams(self);
+        self->flags = FLAG_POS_CAMERA_LOCKED | FLAG_KEEP_ALIVE_OFFCAMERA |
+                      FLAG_HAS_PRIMS | FLAG_UNK_20000;
+        line = (PrimLineG2*)&g_PrimBuf[self->primIndex];
+        self->facingLeft = 0;
+        while (line != NULL) {
+            line->r0 = 0x1F;
+            line->g0 = 0x1F;
+            line->b0 = 0x30;
+            line->r1 = 0x3F;
+            line->g1 = 0x50;
+            line->b1 = 0x7F;
+            line->x1 = line->x0 = rand() & 0x1FF;
+            line->y0 = line->y1 = -(rand() & 0xF);
+            line->preciseX.i.hi = line->x1;
+            line->preciseY.i.hi = line->y1;
+
+            // This whole block is weird. Why are we calculating rcos
+            // and rsin on a fixed value at runtime? And why aren't
+            // these simple multiplications?
+            trigresult = rcos(0xB80);
+            trigtemp = trigresult * 16;
+            line->velocityX.val = (trigresult * 32 + trigtemp) * 4;
+            trigresult = rsin(0xB80);
+            trigtemp = trigresult * -16;
+            line->velocityY.val = trigtemp * 12;
+
+            line->timer = 0;
+            line->delay = (rand() & 0xF) + 12;
+            if (rand() & 1) {
+                line->priority = PLAYER.zPriority + 2;
+            } else {
+                line->priority = PLAYER.zPriority - 2;
+            }
+            line->drawMode = DRAW_TPAGE2 | DRAW_TPAGE | DRAW_TRANSP;
+            line = line->next;
+        }
+        if (self->params == 1) {
+            g_api.func_801027C4(3);
+        }
+        self->ext.subweapon.timer = 0x160;
+        if ((self->params < 32) && !(self->params & 3)) {
+            g_api.PlaySfx(SFX_RIC_HYDRO_STORM_ATTACK);
+        }
+        self->step++;
+        break;
+    case 1:
+        line = (PrimLineG2*)&g_PrimBuf[self->primIndex];
+        while (line != NULL) {
+            if (line->timer == 0) {
+                line->preciseX.i.hi = line->x1;
+                line->preciseY.i.hi = line->y1;
+                line->preciseX.val += line->velocityX.val;
+                line->preciseY.val += line->velocityY.val;
+                line->x1 = line->preciseX.i.hi;
+                line->y1 = line->preciseY.i.hi;
+                if (line->delay < line->y1) {
+                    line->timer++;
+                    line->xLength = line->x0 - line->x1;
+                    line->yLength = line->y0 - line->y1;
+                }
+            } else {
+                line->preciseX.i.hi = line->x1;
+                line->preciseY.i.hi = line->y1;
+                line->preciseX.val += line->velocityX.val;
+                line->preciseY.val += line->velocityY.val;
+                line->x1 = line->preciseX.i.hi;
+                line->y1 = line->preciseY.i.hi;
+                line->y0 = line->y1 + line->yLength;
+                line->x0 = line->x1 + line->xLength;
+                if (line->y0 >= 0xD8) {
+                    self->step = 2;
+                }
+            }
+            line = line->next;
+        }
+        self->ext.subweapon.timer++;
+        break;
+    case 2:
+#if defined(VERSION_HD)
+        if (self->params == 0x28) {
+            g_Player.unk4E = 1;
+        }
+#endif
+        DestroyEntity(self);
+        break;
+    }
+    g_Player.timers[PL_T_3] = 16;
+}
+
+// Copy of DRA function
+s32 RicCheckHolyWaterCollision(s32 baseY, s32 baseX) {
+    s16 x;
+    s16 y;
+    Collider res1;
+    Collider res2;
+    s16 colRes1;
+    s16 colRes2;
+
+    const u32 colFullSet =
+        (EFFECT_UNK_8000 | EFFECT_UNK_4000 | EFFECT_UNK_2000 | EFFECT_UNK_1000 |
+         EFFECT_UNK_0800 | EFFECT_SOLID);
+    const u32 colSetNo800 = (EFFECT_UNK_8000 | EFFECT_UNK_4000 |
+                             EFFECT_UNK_2000 | EFFECT_UNK_1000 | EFFECT_SOLID);
+    const u32 colSet1 = (EFFECT_UNK_8000 | EFFECT_UNK_0800 | EFFECT_SOLID);
+    const u32 colSet2 = (EFFECT_UNK_8000 | EFFECT_SOLID);
+    x = baseX + g_CurrentEntity->posX.i.hi;
+    y = baseY + g_CurrentEntity->posY.i.hi;
+
+    g_api.CheckCollision(x, y, &res1, 0);
+    colRes1 = res1.effects & colFullSet;
+    g_api.CheckCollision(x, (s16)(y - 1 + res1.unk18), &res2, 0);
+    y = baseY + (g_CurrentEntity->posY.i.hi + res1.unk18);
+
+    if ((colRes1 & colSet1) == EFFECT_SOLID ||
+        (colRes1 & colSet1) == (EFFECT_UNK_0800 | EFFECT_SOLID)) {
+        colRes2 = res2.effects & colSetNo800;
+        if (!((s16)res2.effects & EFFECT_SOLID)) {
+            g_CurrentEntity->posY.i.hi = y;
+            return 1;
+        }
+        if ((res2.effects & colSet2) == colSet2) {
+            g_CurrentEntity->posY.i.hi = y + (s16)(res2.unk18 - 1);
+            return colRes2;
+        }
+    } else if ((colRes1 & colSet2) == colSet2) {
+        g_CurrentEntity->posY.i.hi = y;
+        return colRes1 & colSetNo800;
+    }
+    return 0;
+}
+
+s32 func_8016840C(s16 x, s16 y) {
+    Collider collider;
+    u16 temp;
+
+    if (g_CurrentEntity->velocityX != 0) {
+        g_api.CheckCollision(g_CurrentEntity->posX.i.hi + y,
+                             g_CurrentEntity->posY.i.hi + x, &collider, 0);
+        if (g_CurrentEntity->velocityX > 0) {
+            temp = collider.unk14;
+        } else {
+            temp = collider.unk1C;
+        }
+        if (!(collider.effects & EFFECT_UNK_0002)) {
+            return 0;
+        }
+    } else {
+        return 0;
+    }
+    g_CurrentEntity->posX.i.lo = 0;
+    g_CurrentEntity->posX.i.hi += temp;
+    return 2;
+}
+
+// Unused?
+static s16 D_80155D94[] = {16, 0, -1, 0};
+
+// Entity ID #7. Made by blueprint 6. Comes from subweapon 3. Holy water!
+// Not at all the same as DRA's.
+void RicEntitySubwpnHolyWater(Entity* self) {
+    s16 argY;
+    s32 collision_result;
+    s32 tempXVel;
+    s32 trigresult;
+    s32 trigtemp;
+
+    if (self->step >= 3) {
+        self->posY.i.hi += 5;
+    }
+    switch (self->step) {
+    case 0:
+        self->flags = FLAG_POS_CAMERA_LOCKED;
+        self->animSet = ANIMSET_OVL(0x12);
+        self->animCurFrame = 0x23;
+        self->zPriority = PLAYER.zPriority + 2;
+        self->unk5A = 0x46;
+        self->palette = PAL_OVL(0x12F);
+        self->posX.i.hi = self->posX.i.hi;
+        self->posY.i.hi = self->posY.i.hi - 0x10;
+        self->ext.holywater.angle = (rand() & 0x7F) + 0xDC0;
+        if (PLAYER.facingLeft == 1) {
+            self->ext.holywater.angle = (rand() & 0x7F) + 0x9C0;
+        }
+
+        trigresult = rcos(self->ext.holywater.angle);
+        trigtemp = trigresult * 16;
+        self->velocityX = (trigresult * 32 + trigtemp) << 9 >> 8;
+
+        trigresult = rsin(self->ext.holywater.angle);
+        trigtemp = trigresult * 16;
+        self->velocityY = -((trigresult * 32 + trigtemp) << 9) >> 8;
+
+        self->ext.holywater.subweaponId = PL_W_HOLYWATER;
+        RicSetSubweaponParams(self);
+        self->hitboxWidth = 4;
+        self->hitboxHeight = 4;
+        self->ext.holywater.unk80 = 0x200;
+        self->step += 1;
+        break;
+    case 1:
+        self->posY.val += self->velocityY;
+        collision_result = RicCheckHolyWaterCollision(0, 0);
+        self->posX.val += self->velocityX;
+
+        if (self->velocityX < 0) {
+            argY = -4;
+        } else {
+            argY = 4;
+        }
+        collision_result |= func_8016840C(-7, argY);
+        if (collision_result & 2) {
+            self->velocityX = -self->velocityX;
+            collision_result = 1;
+        }
+        if ((collision_result & 1) || (self->hitFlags != 0)) {
+            g_api.CreateEntFactoryFromEntity(self, 0x28U, 0);
+            g_api.PlaySfx(SFX_GLASS_BREAK_WHOOSH);
+            self->ext.holywater.timer = 0x50;
+            self->animSet = 0;
+            self->hitboxState = 0;
+            self->step = 3;
+            self->velocityX >>= 2;
+        }
+        break;
+    case 2:
+        if (--self->ext.holywater.timer == 0) {
+            self->ext.holywater.timer = 0x50;
+            self->velocityX >>= 2;
+            self->step++;
+        }
+        break;
+    case 3:
+        if (!(self->ext.holywater.timer & 3)) {
+            RicCreateEntFactoryFromEntity(
+                self, FACTORY(BP_HOLYWATER_FIRE, self->ext.holywater.unk82), 0);
+            // weirdly nonmatching with ++;
+            self->ext.holywater.unk82 = self->ext.holywater.unk82 + 1;
+            self->velocityX -= (self->velocityX / 32);
+        }
+        collision_result = RicCheckHolyWaterCollision(6, 0);
+        tempXVel = self->velocityX;
+        if (self->velocityX < 0) {
+            if ((collision_result & 0xF000) == 0xC000) {
+                tempXVel = tempXVel * 10 / 16;
+            }
+            if ((collision_result & 0xF000) == 0xD000) {
+                tempXVel = tempXVel * 13 / 16;
+            }
+        }
+        if (self->velocityX > 0) {
+            if ((collision_result & 0xF000) == 0x8000) {
+                tempXVel = tempXVel * 10 / 16;
+            }
+            if ((collision_result & 0xF000) == 0x9000) {
+                tempXVel = tempXVel * 13 / 16;
+            }
+        }
+        self->posX.val += tempXVel;
+        if (self->velocityX < 0) {
+            argY = -4;
+        } else {
+            argY = 4;
+        }
+        collision_result |= func_8016840C(-7, argY);
+        if (!(collision_result & 1)) {
+            self->velocityX >>= 1;
+            self->step++;
+        }
+        break;
+    case 4:
+        if (!(self->ext.holywater.timer & 3)) {
+            RicCreateEntFactoryFromEntity(
+                self, FACTORY(BP_HOLYWATER_FIRE, self->ext.holywater.unk82), 0);
+            self->ext.holywater.unk82 = self->ext.holywater.unk82 + 1;
+        }
+        self->velocityY += FIX(12.0 / 128);
+        if (self->velocityY > FIX(4)) {
+            self->velocityY = FIX(4);
+        }
+        self->posY.val += self->velocityY;
+        collision_result = RicCheckHolyWaterCollision(0, 0);
+        self->posX.val += self->velocityX;
+        if (self->velocityX < 0) {
+            argY = -4;
+        } else {
+            argY = 4;
+        }
+        collision_result |= func_8016840C(-7, argY);
+        if (collision_result & 1) {
+            self->velocityX *= 2;
+            self->step--;
+        }
+        break;
+    case 5:
+        break;
+    }
+    if (self->step >= 3) {
+        if (--self->ext.holywater.timer < 0) {
+            DestroyEntity(self);
+            return;
+        }
+        if (self->ext.holywater.timer == 2) {
+            self->step = 5;
+        }
+        self->animCurFrame = 0;
+        self->hitboxState = 0;
+        self->posY.i.hi -= 5;
+    }
+    g_Player.timers[PL_T_3] = 2;
+}
+
+// Entity ID #8. Blueprint 7.
+static Point32 D_80155D9C[] = {{28, 0}, {28, 16}, {28, 32}, {28, 48},
+                               {60, 0}, {60, 16}, {60, 32}, {60, 48}};
+void RicEntitySubwpnHolyWaterFlame(Entity* self) {
+    s16 sp10[5];
+    s16 sp20[5];
+    s16 pad[2];
+    Primitive* prim;
+    s16 angleTemp;
+    s32 angle;
+    s32 i;
+    s32 hex80;
+    u8 randR;
+    u8 randG;
+    u8 randB;
+    u8 primUBase;
+    u8 primVBase;
+    s16* primYPtr;
+    s32 var_s4;
+    s16 temp_v0_2;
+    s32 doubleparams;
+
+    s16 upperParams = self->params >> 8;
+
+    primUBase = D_80155D9C[(g_GameTimer & 7)].x;
+    primVBase = D_80155D9C[(g_GameTimer & 7)].y;
+    switch (self->step) {
+    case 0:
+        randR = (rand() & 0x3F) + 0x2F;
+        randG = (rand() & 0x3F) + 0x6F;
+        randB = (rand() & 0x7F) + 0x6F;
+        self->primIndex = g_api.AllocPrimitives(PRIM_GT4, 4);
+        if (self->primIndex == -1) {
+            DestroyEntity(self);
+            return;
+        }
+        prim = &g_PrimBuf[self->primIndex];
+        i = 0;
+        while (prim != NULL) {
+            PRED(prim) = randR;
+            PGRN(prim) = randG;
+            PBLU(prim) = randB;
+            if (i == 0) {
+                prim->b0 = prim->b1 = prim->g0 = prim->g1 = prim->r0 =
+                    prim->r1 = 0;
+            }
+            prim->clut = 0x1B0;
+            prim->tpage = 0x1A;
+            prim->priority = PLAYER.zPriority + 2;
+            prim->drawMode = DRAW_UNK_40 | DRAW_TPAGE2 | DRAW_TPAGE |
+                             DRAW_HIDE | DRAW_COLORS | DRAW_UNK02 | DRAW_TRANSP;
+            prim = prim->next;
+            i++;
+        }
+        self->flags = FLAG_POS_CAMERA_LOCKED | FLAG_HAS_PRIMS;
+        self->ext.holywater.timer = 1;
+        self->step += 1;
+        break;
+    case 1:
+        if (--self->ext.holywater.timer == 0) {
+            self->ext.holywater.subweaponId = PL_W_HOLYWATER_FLAMES;
+            RicSetSubweaponParams(self);
+            self->hitboxWidth = 4;
+            self->ext.holywater.hitboxState = self->hitboxState;
+            self->posY.i.hi = self->posY.i.hi - 10;
+            RicCreateEntFactoryFromEntity(self, FACTORY(BP_EMBERS, 3), 0);
+            self->ext.holywater.timer = 0x50;
+            self->posY.i.hi = self->posY.i.hi + 10;
+            self->ext.holywater.unk80 = (rand() & 0xF) + 0x12;
+            self->ext.holywater.angle = rand() & 0xFFF;
+            self->step += 1;
+        }
+        break;
+    case 2:
+        if (self->facingLeft) {
+            var_s4 = 1;
+        } else {
+            var_s4 = -1;
+        }
+        angleTemp = self->ext.holywater.angle;
+        self->ext.holywater.angle += 0xC0;
+        angle = angleTemp;
+        for (i = 0; i < 4; i++) {
+            sp10[i] = self->posX.i.hi + (rsin(angle) >> 0xA);
+            angle += 0x400;
+        }
+        sp10[4] = self->posX.i.hi;
+        sp10[0] = var_s4 + self->posX.i.hi;
+        temp_v0_2 = rsin((s16)((self->ext.holywater.timer * 64) + 0x800)) >> 8;
+        temp_v0_2 += self->ext.holywater.unk80;
+        temp_v0_2 = temp_v0_2 * 3 >> 1;
+        sp20[0] = self->posY.i.hi - temp_v0_2;
+        sp20[4] = self->posY.i.hi;
+        sp20[2] = (sp20[0] + sp20[4]) / 2;
+        sp20[1] = (sp20[0] + sp20[2]) / 2;
+        sp20[3] = (sp20[2] + sp20[4]) / 2;
+        prim = &g_PrimBuf[self->primIndex];
+        if (--self->ext.holywater.timer < 0) {
+            DestroyEntity(self);
+            return;
+        }
+        if (self->ext.holywater.timer & 3) {
+            self->hitboxState = 0;
+        } else {
+            self->hitboxState = self->ext.holywater.hitboxState;
+        }
+        if (self->ext.holywater.timer < 0x15) {
+            self->hitboxState = 0;
+        }
+        i = 0;
+        while (prim != NULL) {
+            if (upperParams * 2 + 0x18 >= self->ext.holywater.timer) {
+                if (prim->g0 >= 10) {
+                    prim->g0 -= 5;
+                }
+                if (prim->b0 >= 10) {
+                    prim->b0 -= 5;
+                }
+                if (prim->r0 >= 16) {
+                    prim->r0 -= 8;
+                }
+                prim->g1 = prim->g2 = prim->g3 = prim->g0;
+                prim->r1 = prim->r2 = prim->r3 = prim->r0;
+                prim->b1 = prim->b2 = prim->b3 = prim->b0;
+            }
+            prim->x0 = sp10[i] - 8;
+            prim->x1 = sp10[i] + 8;
+            prim->y0 = sp20[i];
+            prim->y1 = sp20[i];
+            prim->x2 = sp10[i + 1] - 8;
+            prim->x3 = sp10[i + 1] + 8;
+            prim->y2 = sp20[i + 1];
+            prim->y3 = sp20[i + 1];
+            prim->drawMode &= ~DRAW_HIDE;
+            hex80 = 0x80;
+            prim->u0 = prim->u1 = primUBase - ((i * 7) + hex80);
+            prim->u2 = prim->u3 = primUBase - (((i + 1) * 7) + hex80);
+            prim->v0 = prim->v2 = primVBase - hex80;
+            prim->v1 = prim->v3 = primVBase - 0x70;
+            if ((sp20[4] - sp20[0]) < 7) {
+                prim->drawMode |= DRAW_HIDE;
+                DestroyEntity(self);
+                return;
+            }
+            i++;
+            prim = prim->next;
+        }
+        self->hitboxHeight = temp_v0_2 >> 1;
+        self->hitboxOffY = (-temp_v0_2 >> 1);
+    }
+    g_Player.timers[PL_T_3] = 2;
+}
